requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sentence-transformers==2.6.1
faiss-cpu==1.7.2
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.31.0
